switched noise to 4 channel bgra upload

+++ project-griffin/project-griffin/source/render/impl/RenderTarget_GL.cpp
@@ -64,7 +64,7 @@ namespace griffin {
 				glGenTextures(1, &albedo);
 				m_textureIds[Albedo_Displacement] = albedo;
 				glBindTexture(GL_TEXTURE_2D, albedo);
-				glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA8, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
+				glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
